refactor(ui): split imgui_state.h from imgui_ui.h — P1.5.2 step 1
`imgui_ui.h` previously mixed the data side of the UI (the
`ImGuiUIState` singleton, telemetry fields the core writes into, file
dialog enum) with declarations of UI-only functions (`imgui_init_ui`,
`imgui_render_ui`, `imgui_topbar_height`, toasts, …).

For the headless build target (P1.5.2) the core translation units need
access to `imgui_state` but must not see UI function declarations,
otherwise the linker fails to resolve them in the MODERN_UI=OFF build
where `imgui_ui.cpp` / `devtools_ui.cpp` / `video.cpp` are excluded.

Extract the headless-safe parts into a new `src/imgui_state.h`:
`FileDialogAction`, `ImGuiUIState`, the `extern` declaration of the
singleton.  Move the singleton's definition into its own TU
`src/imgui_state.cpp` so it's part of `koncepcja_lib` regardless of
the MODERN_UI flag.  `imgui_ui.h` now `#include`s the new header so
existing modern-UI callers keep compiling unchanged.

Switch the four headless-only consumers
(`koncepcja_ipc_server.cpp`, `tape.cpp`, `z80.cpp`, `test/tape.cpp`)
to include `imgui_state.h` instead of `imgui_ui.h`, demonstrating the
split actually works.  `test/ui_state_test.cpp` keeps `imgui_ui.h`
because it calls UI functions (close_menu, toast_info).

Pure refactor, no behaviour change.  1114/1114 unit tests pass, same
two pre-existing RamExpansion skips.  Build is bit-identical with
MODERN_UI=ON.  MODERN_UI=OFF link still fails (will be fixed in step
2 by stubbing the remaining UI calls in `kon_cpc_ja.cpp`'s main loop).

bd: beads-6oa
